Remove mirror_*_download functions
mirror_target_download and mirror_meta_download functions that
return iterators are causing more troubles than good.

Since there is a need to download and verify each file inside
the mirrors loop (before continuing to the next mirror), feels
like the correct component to do this is the Updater itself.
This means slightly repetitive code inside the Updater class
but with the benefit of better readability and less bug-prone
implementation.

Signed-off-by: Teodora Sechkova <tsechkova@vmware.com>